Extract event components
