+++ ui-components/src/page_components/listing/listing_sidebar/events/LotteryResultsEvent.tsx
@@ -1,0 +1,26 @@
+import * as React from "react"
+import { ListingEvent } from "@bloom-housing/backend-core/types"
+import { t } from "../../../../helpers/translator"
+import moment from "moment"
+
+const LotteryResultsEvent = (props: { event: ListingEvent }) => {
+  const { event } = props
+  return (
+    <section className="aside-block -mx-4 pt-0 md:mx-0 md:pt-4">
+      <h4 className="text-caps-underline">{t("listings.lotteryResults.header")}</h4>
+      <p className="text text-gray-800 pb-3 flex justify-between items-center">
+        <span className="inline-block">{moment(props.event.startTime).format("MMMM D, YYYY")}</span>
+      </p>
+      {event.note && <p className="text text-gray-600">{event.note}</p>}
+      {!event.note && (
+        <p className="text text-gray-600">
+          {t("listings.lotteryResults.completeResultsWillBePosted", {
+            hour: moment(props.event.startTime).format("h a"),
+          })}
+        </p>
+      )}
+    </section>
+  )
+}
+
+export { LotteryResultsEvent as default, LotteryResultsEvent }
