+++ ui-components/src/page_components/listing/listing_sidebar/events/OpenHouseEvent.tsx
@@ -1,0 +1,25 @@
+import * as React from "react"
+import { ListingEvent } from "@bloom-housing/backend-core/types"
+import { EventDateSection } from "../../../../sections/EventDateSection"
+import { t } from "../../../../helpers/translator"
+
+const OpenHouseEvent = (props: { openHouseEvents: ListingEvent[] }) => {
+  return (
+    <section className="aside-block bg-primary-lighter border-t">
+      <h4 className="text-caps-tiny">{t("listings.openHouseEvent.header")}</h4>
+      {props.openHouseEvents.map((openHouseEvent, index) => (
+        <div key={`openHouses-${index}`}>
+          <EventDateSection event={openHouseEvent} />
+          {openHouseEvent.url && (
+            <p className="text text-gray-800 pb-3">
+              <a href={openHouseEvent.url}>{t("listings.openHouseEvent.seeVideo")}</a>
+            </p>
+          )}
+          {openHouseEvent.note && <p className="text text-gray-600">{openHouseEvent.note}</p>}
+        </div>
+      ))}
+    </section>
+  )
+}
+
+export { OpenHouseEvent as default, OpenHouseEvent }
