+++ ui-components/src/sections/EventDateSection.tsx
@@ -1,0 +1,18 @@
+import * as React from "react"
+import { ListingEvent } from "@bloom-housing/backend-core/types"
+import moment from "moment"
+
+const EventDateSection = (props: { event: ListingEvent }) => {
+  return (
+    <p className="text text-gray-800 pb-3 flex justify-between items-center">
+      <span className="inline-block">{moment(props.event.startTime).format("MMMM D, YYYY")}</span>
+      <span className="inline-block text-xs font-bold">
+        {moment(props.event.startTime).format("hh:mma") +
+          "-" +
+          moment(props.event.endTime).format("hh:mma")}
+      </span>
+    </p>
+  )
+}
+
+export { EventDateSection as default, EventDateSection }
